refactor(PIO-49): fix convention warnings — return type and test file size
- Add explicit return type ?Subscription to User::getSubscriptionAttribute()
- Split CliAuthTest.php (715 lines) into three focused files, each under 500 lines:
  - CliAuthTest.php: page rendering and token exchange tests
  - CliDeviceNameTest.php: device name and re-authorisation tests
  - MobileAuthTest.php: mobile OAuth flow tests
